#121 TestJson: refactor JSON tests to use common testset for both reader and writer
