+++ tlair/src/test/scala/at/forsyte/apalache/tla/lir/io/TestJsonReader.scala
@@ -1,24 +1,11 @@
 package at.forsyte.apalache.tla.lir.io
 
-import at.forsyte.apalache.tla.lir.convenience.tla._
 import at.forsyte.apalache.tla.lir._
-import org.junit.runner.RunWith
-import org.scalatest.junit.JUnitRunner
-import org.scalatest.FunSuite
 
-@RunWith(classOf[JUnitRunner])
-class TestJsonReader extends FunSuite {
+class TestJsonReader extends TestJson {
 
-  // compare expression and expected result (single-line formatting)
-  def compare(expected: TlaEx, json: String): Unit = {
+  // check that after reading from json the result matches expr
+  def compare(expected: TlaEx, json: String, indent: Int = -1): Unit = {
     assert(JsonReader.read(json) == expected)
-  }
-
-  test("str") {
-    // "Hello, World!"
-    compare(
-      str("Hello, World!"),
-      """{"str":"Hello, World!"}"""
-    )
   }
 }
